+++ NEWS.md
@@ -1,3 +1,12 @@
+## version 1.2.0
+- March-2025
+- Added methods for lpda and lpdaCV: print and summary
+- Previous CV functions grouped in lpdaCV function
+- Added functions to deal with 3dimensional arrays
+- RNAseq data updated for a 3dimensional array
+- Vignette and help updated
+
+
 ## version 1.0.1
 
 - March-2023
@@ -13,6 +22,7 @@
 
 - March-2023
 - Updated Vignette
+